implement more intrinsics
